# GeometryCollection: Cache flags to avoid O(n) lookups