--- a/benchmarks/operation/CMakeLists.txt
+++ b/benchmarks/operation/CMakeLists.txt
@@ -19,4 +19,12 @@ if (benchmark_FOUND)
             $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/benchmarks>)
     target_link_libraries(perf_distance PRIVATE
             benchmark::benchmark geos geos_cxx_flags)
+
+    add_executable(perf_coverage_union CoverageUnionPerfTest.cpp)
+    target_include_directories(perf_coverage_union PUBLIC
+            $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
+            $<BUILD_INTERFACE:${PROJECT_BINARY_DIR}/include>
+            $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/benchmarks>)
+    target_link_libraries(perf_coverage_union PRIVATE
+            benchmark::benchmark geos geos_cxx_flags)
 endif()