@dr-jts 2025-01-04
> Increases the size of collections from 88 to 96 bytes.

Seems like a low-percentage increase.